+++ wellregistry/registry/admin/fetch_from_nwis.py
@@ -1,6 +1,7 @@
 """
 Custom fetch from nwis form and view for the Django admin
 """
+
 import requests
 
 from django.conf import settings
@@ -10,7 +11,9 @@ from django.shortcuts import render, redirect
 from django.urls.base import reverse
 from django.views.generic.edit import FormView
 
-from ..models import MonitoringLocation, AgencyLookup, AltitudeDatumLookup, HorizontalDatumLookup, \
+from wellregistry import nwis_aquifer_lookups
+
+from ..models import MonitoringLocation, AgencyLookup, AltitudeDatumLookup, UnitsLookup, HorizontalDatumLookup, \
     NatAqfrLookup, CountryLookup, StateLookup, CountyLookup
 from ..utils import parse_rdb
 
@@ -42,6 +45,13 @@ class FetchFromNwisView(FormView):
         if not site_data['well_depth_va']:
             return False, 'Site is missing a well depth'
         return True, 'Valid site'
+
+    @staticmethod
+    def _get_local_aquifer_name(list_aqr_lookups, aqfr_cd, state_cd):
+        for item in list_aqr_lookups:
+            if item['Aqfr_Cd'] == aqfr_cd and item['State_Cd'] == state_cd:
+                return item['Aqfr_Nm']
+        raise KeyError
 
     @staticmethod
     def _get_monitoring_location(site_data, user):
@@ -89,10 +99,14 @@ class FetchFromNwisView(FormView):
         monitoring_location.horz_acy = site_data['coord_acy_cd']
         monitoring_location.alt_va = float(site_data['alt_va'])
         monitoring_location.altitude_datum = AltitudeDatumLookup.objects.get(adatum_cd=site_data['alt_datum_cd'])
+        monitoring_location.altitude_units=UnitsLookup.objects.get(unit_id=1)
         monitoring_location.alt_method = site_data['alt_meth_cd']
         monitoring_location.alt_acy = site_data['alt_acy_va']
         monitoring_location.well_depth = float(site_data['well_depth_va'])
+        monitoring_location.well_depth_units = UnitsLookup.objects.get(unit_id=1)
         monitoring_location.nat_aqfr = NatAqfrLookup.objects.get(nat_aqfr_cd=site_data['nat_aqfr_cd'])
+        monitoring_location.local_aquifer_name = \
+            FetchFromNwisView._get_local_aquifer_name(nwis_aquifer_lookups, site_data['aqfr_cd'], site_data['state_cd'])
         monitoring_location.site_type = 'SPRING' if site_data['site_tp_cd'] == 'SP' else 'WELL'
         monitoring_location.aqfr_type = nwis_aqfr_type_cd_to_aqfr_type[site_data['aqfr_type_cd']]
         monitoring_location.update_user = user